Refactor code: Split classes into smaller packages
